% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_effsize.R
\name{compute_effsize}
\alias{compute_effsize}
\title{This function should be the main API}
\usage{
compute_effsize(
  type,
  effsize,
  colname = "yi",
  Ctrl_mean,
  Ctrl_sd,
  Ctrl_n,
  A_mean,
  A_sd,
  A_n,
  B_mean,
  B_sd,
  B_n,
  AB_mean,
  AB_sd,
  AB_n,
  data
)
}
\description{
Needs for each effect size:
}
\details{
lnRR, SMD and Hedges' g: the 12 params
lnVR and lnCVR: only SD and N.
}
