% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator_functions.R
\name{create_event_pulses}
\alias{create_event_pulses}
\title{Create event pulse features}
\usage{
create_event_pulses(abs_t, events, jitter_days)
}
\description{
Create event pulse features
}
\keyword{internal}
