% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator_functions.R
\name{create_scenarios}
\alias{create_scenarios}
\title{Create Scenarios for Emulator}
\usage{
create_scenarios(...)
}
\arguments{
\item{...}{Named parameters with values}
}
\value{
Data frame of scenarios
}
\description{
Create Scenarios for Emulator
}
