% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator_functions.R
\name{create_time_since}
\alias{create_time_since}
\title{Create time-since-event features}
\usage{
create_time_since(abs_t, events)
}
\description{
Create time-since-event features
}
\keyword{internal}
