% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{delay_load_python}
\alias{delay_load_python}
\title{Delay Load Python Modules}
\usage{
delay_load_python()
}
\description{
Sets up delayed loading of Python modules to avoid loading them
until they are actually needed.
}
\keyword{internal}
