% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.compute_and_format}
\alias{.compute_and_format}
\title{Compute And Format}
\usage{
.compute_and_format(effsize_func, effsize_args, data, col_names, append)
}
\arguments{
\item{effsize_func}{Function used to compute the effect size}

\item{effsize_args}{Arguments needed for that function}

\item{data}{Data frame to use}

\item{col_names}{Vector of length 2 with the name of the columns}

\item{append}{Logical. If append or not the effect sizes to the data}
}
\description{
Compute effect size and do some pre-return tasks
}
\keyword{internal}
