% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedges_g.R
\name{.compute_smd}
\alias{.compute_smd}
\title{Simple effect: Standardized mean difference Hedges' g}
\usage{
.compute_smd(
  mean_control,
  mean_treatment,
  sd_control,
  sd_treatment,
  n_control,
  n_treatment
)
}
\description{
Simple effect: Standardized mean difference Hedges' g
}
