% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{.compute_var_smd}
\alias{.compute_var_smd}
\title{Computes sampling variance for Standardized Mean Difference}
\usage{
.compute_var_smd(d, Ctrl_n, X_n)
}
\arguments{
\item{d}{Mean difference previously calculated}

\item{Ctrl_n}{Sample size from the control streatment}

\item{X_n}{Sample size from treatment}
}
\description{
Computes sampling variance for Standardized Mean Difference
}
\keyword{internal}
