% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnVR.R
\name{.interaction_lnVR}
\alias{.interaction_lnVR}
\title{Interaction effect: Log Variability Ratio}
\usage{
.interaction_lnVR(Ctrl_sd, Ctrl_n, A_sd, A_n, B_sd, B_n, AB_sd, AB_n)
}
\arguments{
\item{Ctrl_sd}{Standard deviation from the control treatment}

\item{Ctrl_n}{Sample size from the control streatment}

\item{A_sd}{Standard deviation from the A treatment}

\item{A_n}{Sample size from the A treatment}

\item{B_sd}{Standard deviation from the B treatment}

\item{B_n}{Sample size from the B treatment}

\item{AB_sd}{Standard deviation from the interaction AxB treatment}

\item{AB_n}{Sample size from the interaction AxB treatment}
}
\description{
Interaction effect: Log Variability Ratio
}
\keyword{internal}
