% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{.j_correction}
\alias{.j_correction}
\title{Correction for small-sample bias}
\usage{
.j_correction(m)
}
\arguments{
\item{m}{Degrees of freedom. They change between individual and main effects.}
}
\description{
Correction for small-sample bias
}
\keyword{internal}
