% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnCVR_wrappers.R
\name{.lnCVR}
\alias{.lnCVR}
\title{Log Coefficient Of Variation Ratio (lnCVR)}
\usage{
.lnCVR(
  type,
  data,
  col_names = c("yi", "vi"),
  append = TRUE,
  Ctrl_mean,
  Ctrl_sd,
  Ctrl_n,
  A_mean,
  A_sd,
  A_n,
  B_mean = NULL,
  B_sd = NULL,
  B_n = NULL,
  AB_mean = NULL,
  AB_sd = NULL,
  AB_n = NULL
)
}
\description{
Log Coefficient Of Variation Ratio (lnCVR)
}
\author{
Facundo Decunta - fdecunta@agro.uba.ar
}
\keyword{internal}
