% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnCVR_wrappers.R
\docType{data}
\name{.lnCVR_requirements}
\alias{.lnCVR_requirements}
\title{Columns required for computing each lnCVR}
\format{
An object of class \code{list} of length 2.
}
\usage{
.lnCVR_requirements
}
\description{
There is no 'inter' because uses the same args than 'main'
}
\keyword{internal}
