% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnRR_wrappers.R
\docType{data}
\name{.lnRR_requirements}
\alias{.lnRR_requirements}
\title{Columns required for computing different lnRR effect sizes}
\format{
An object of class \code{list} of length 2.
}
\usage{
.lnRR_requirements
}
\description{
Columns required for computing different lnRR effect sizes
}
\keyword{internal}
