% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnVR.R
\docType{data}
\name{.lnVR_args}
\alias{.lnVR_args}
\title{Required columns for computing lnVR}
\format{
An object of class \code{list} of length 2.
}
\usage{
.lnVR_args
}
\description{
Required columns for computing lnVR
}
\keyword{internal}
