% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_effsize.R
\name{.name_columns}
\alias{.name_columns}
\title{Rename columns with factor names}
\usage{
.name_columns(effsize, df, factor_names)
}
\arguments{
\item{factor_names}{}
}
\description{
Rename columns with factor names
}
\keyword{internal}
