% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnCVR.R
\name{.simple_lnCVR}
\alias{.simple_lnCVR}
\title{Simple Log Coefficient Of Variation Ratio}
\usage{
.simple_lnCVR(Ctrl_mean, Ctrl_sd, Ctrl_n, A_mean, A_sd, A_n)
}
\arguments{
\item{Ctrl_mean}{Mean outcome from the Control treatment}

\item{Ctrl_sd}{Standard deviation from the control treatment}

\item{Ctrl_n}{Sample size from the control streatment}

\item{A_mean}{Mean outcome from the experimental treatment}

\item{A_sd}{Standard deviation from the experimental treatment}

\item{A_n}{Sample size from the experimental treatment}
}
\description{
TODO:
}
\references{
Nakagawa, S., Poulin, R., Mengersen, K., Reinhold, K., Engqvist,
L., Lagisz, M., & Senior, A. M. (2015). Meta‐analysis of variation:
ecological and evolutionary applications and beyond. Methods in
Ecology and Evolution, 6(2), 143-152.
}
\keyword{internal}
