% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnRR.R
\name{.simple_lnRR}
\alias{.simple_lnRR}
\title{Simple Log Response Ratio}
\usage{
.simple_lnRR(Ctrl_mean, Ctrl_sd, Ctrl_n, A_mean, A_sd, A_n)
}
\arguments{
\item{Ctrl_mean}{Mean outcome from the Control treatment}

\item{Ctrl_sd}{Standard deviation from the control treatment}

\item{Ctrl_n}{Sample size from the control streatment}

\item{A_mean}{Mean outcome from the treatment}

\item{A_sd}{Standard deviation from the treatment}

\item{A_n}{Sample size from the treatment}
}
\description{
Computes the log of the response ratio
}
\references{
Morris, W. F., Hufbauer, R. A., Agrawal, A. A., Bever, J. D., Borowicz, V. A.,
Gilbert, G. S., ... & Vázquez, D. P. (2007). Direct and interactive
effects of enemies and mutualists on plant performance: a meta‐analysis.
Ecology, 88(4), 1021-1029. https://doi.org/10.1890/06-0442

Lajeunesse, M. J. (2011). On the meta‐analysis of response ratios for
studies with correlated and multi‐group designs. Ecology, 92(11), 2049-2055.
https://doi.org/10.1890/11-0423.1
}
\keyword{internal}
