% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMD_wrappers.R
\name{.SMD}
\alias{.SMD}
\title{Standardized Mean Difference (SMD)}
\usage{
.SMD(
  type,
  data,
  col_names = c("yi", "vi"),
  append = TRUE,
  hedges_correction = TRUE,
  Ctrl_mean,
  Ctrl_sd,
  Ctrl_n,
  A_mean,
  A_sd,
  A_n,
  B_mean = NULL,
  B_sd = NULL,
  B_n = NULL,
  AB_mean = NULL,
  AB_sd = NULL,
  AB_n = NULL
)
}
\arguments{
\item{hedges_correction}{Logical. Apply or not Hedges' correction for small-sample bias. Default is TRUE}
}
\description{
Standardized Mean Difference (SMD)
}
\author{
Facundo Decunta - fdecunta@agro.uba.ar
}
\keyword{internal}
