% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_SMD.R
\name{.time_pooled_sd}
\alias{.time_pooled_sd}
\title{Pooled Standard Deviation for SMD in non-independent factorial}
\usage{
.time_pooled_sd(t0_Ctrl_sd, t1_Ctrl_sd, Ctrl_n, t0_Exp_sd, t1_Exp_sd, Exp_n)
}
\arguments{
\item{t0_Ctrl_sd}{Standard deviation from Control treatment at time 0}

\item{t1_Ctrl_sd}{Standard deviation from Control treatment at time 1}

\item{Ctrl_n}{Sample size from Control treatment}

\item{t0_Exp_sd}{Standard deviation from Experimental treatment at time 0}

\item{t1_Exp_sd}{Standard deviation from Experimental treatment at time 1}

\item{Exp_n}{Sample size from Experimental treatment}
}
\description{
Pooled Standard Deviation for SMD in non-independent factorial
}
\keyword{internal}
