% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_mod_results.R
\name{factorial_mod_results}
\alias{factorial_mod_results}
\title{Aggregate the results from multiple models}
\usage{
factorial_mod_results(models, group)
}
\arguments{
\item{models}{A named list with the models and their name.}

\item{group}{Grouping variable. See \code{orchard::group}}
}
\description{
Aggregate the results from multiple models
}
