% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_data}
\alias{fake_data}
\title{Example meta-analysis dataset}
\format{
A data frame with 9 rows and 10 columns.
}
\source{
Simulated
}
\usage{
fake_data
}
\description{
A dummy dataset with means, SDs, and sample sizes for factorial meta-analysis.
}
\keyword{datasets}
