% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator_functions.R
\name{generate_scenario_predictions}
\alias{generate_scenario_predictions}
\title{Generate Scenario Predictions}
\usage{
generate_scenario_predictions(
  scenarios,
  models,
  model_types = c("GRU", "LSTM"),
  time_steps = 2190
)
}
\arguments{
\item{scenarios}{Data frame with scenario parameters}

\item{models}{List from load_emulator_models()}

\item{model_types}{Vector of model types to use ("GRU", "LSTM", or both)}

\item{time_steps}{Number of time steps to predict (in days for prevalence)}
}
\value{
List with predictions for each scenario
}
\description{
Generate Scenario Predictions
}
