% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{get_runtime_parameters}
\alias{get_runtime_parameters}
\title{Get Runtime Parameters for Simulation}
\usage{
get_runtime_parameters(
  i,
  lhs_data,
  HUMAN_POPULATION,
  bednet_params,
  SIM_LENGTH
)
}
\arguments{
\item{i}{Index of parameter set}

\item{lhs_data}{Data frame of LHS scenarios}

\item{HUMAN_POPULATION}{Human population size}

\item{bednet_params}{Bednet parameter data}

\item{SIM_LENGTH}{Simulation length in days}
}
\value{
List of input parameters
}
\description{
Get Runtime Parameters for Simulation
}
