% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{initialize_simulation_parameters}
\alias{initialize_simulation_parameters}
\title{Initialize Simulation Parameters}
\usage{
initialize_simulation_parameters(
  lhs_sample,
  HUMAN_POPULATION,
  selected_seasonality
)
}
\arguments{
\item{lhs_sample}{Single row from LHS data}

\item{HUMAN_POPULATION}{Human population size}

\item{selected_seasonality}{Seasonality parameters}
}
\value{
malariasimulation parameters object
}
\description{
Initialize Simulation Parameters
}
