% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_effsize.R
\name{inter_effsize}
\alias{inter_effsize}
\title{Compute interaction effect sizes}
\usage{
inter_effsize(
  effsize,
  colnames,
  Ctrl_mean,
  Ctrl_sd,
  Ctrl_n,
  A_mean,
  A_sd,
  A_n,
  B_mean,
  B_sd,
  B_n,
  AB_mean,
  AB_sd,
  AB_n,
  data
)
}
\arguments{
\item{effsize}{Character. Effect‐size metric to compute. Supported: \code{"lnrr"}.}

\item{colnames}{Character vector of length 2 giving the names to use for the
factors A and B}

\item{Ctrl_mean}{Mean outcome from the Control treatment}

\item{Ctrl_sd}{Standard deviation from the control treatment}

\item{Ctrl_n}{Sample size from the control streatment}

\item{A_mean}{Mean outcome from the A treatment}

\item{A_sd}{Standard deviation from the A treatment}

\item{A_n}{Sample size from the A treatment}

\item{B_mean}{Mean outcome from the B treatment}

\item{B_sd}{Standard deviation from the B treatment}

\item{B_n}{Sample size from the B treatment}

\item{AB_mean}{Mean outcome from the interaction AxB treatment}

\item{AB_sd}{Standard deviation from the interaction AxB treatment}

\item{AB_n}{Sample size from the interaction AxB treatment}

\item{data}{Data fram with the data.}
}
\value{
The original \code{data} with the effect size and its variance
appended, named as in \code{colnames}.
}
\description{
TODO: !
}
\examples{
\dontrun{
result <- inter_effsize(
  effsize   = "lnrr",
  data      = df,
  colnames  = c("yi", "vi"),
  Ctrl_mean = ctrl_mean, Ctrl_sd = ctrl_sd, Ctrl_n = ctrl_n,
  A_mean    = a_mean,    A_sd    = a_sd,    A_n    = a_n,
  B_mean    = b_mean,    B_sd    = b_sd,    B_n    = b_n,
  AB_mean   = ab_mean,   AB_sd   = ab_sd,   AB_n   = ab_n,
)
}

}
\author{
Facundo Decunta - fdecunta@agro.uba.ar
}
