% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_vcv.R
\name{inter_vcv}
\alias{inter_vcv}
\title{Variance covariance matrices for non-independence}
\usage{
inter_vcv(cols, cluster, obs, rho, data, ...)
}
\description{
Variance covariance matrices for non-independence
}
