% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnrr.R
\name{interaction_lnRR}
\alias{interaction_lnRR}
\title{Interaction effect: log of response ratio}
\usage{
interaction_lnRR(
  Ctrl_mean,
  Ctrl_sd,
  Ctrl_n,
  A_mean,
  A_sd,
  A_n,
  B_mean,
  B_sd,
  B_n,
  AB_mean,
  AB_sd,
  AB_n
)
}
\arguments{
\item{Ctrl_mean}{Mean outcome from the Control treatment}

\item{Ctrl_sd}{Standard deviation from the control treatment}

\item{Ctrl_n}{Sample size from the control streatment}

\item{A_mean}{Mean outcome from the A treatment}

\item{A_sd}{Standard deviation from the A treatment}

\item{A_n}{Sample size from the A treatment}

\item{B_mean}{Mean outcome from the B treatment}

\item{B_sd}{Standard deviation from the B treatment}

\item{B_n}{Sample size from the B treatment}

\item{AB_mean}{Mean outcome from the interaction AxB treatment}

\item{AB_sd}{Standard deviation from the interaction AxB treatment}

\item{AB_n}{Sample size from the interaction AxB treatment}
}
\description{
Computes the interaction effect of a treatment in a 2-by-2 factorial design.
}
\details{
Given a full factorial design with factors A x B
The interacion effect of A x B is quantified as the

TODO: Complete this!!!
}
\references{
Morris, W. F., Hufbauer, R. A., Agrawal, A. A., Bever, J. D., Borowicz, V. A.,
Gilbert, G. S., ... & Vázquez, D. P. (2007). Direct and interactive
effects of enemies and mutualists on plant performance: a meta‐analysis.
Ecology, 88(4), 1021-1029. https://doi.org/10.1890/06-0442
Lajeunesse, M. J. (2011). On the meta‐analysis of response ratios for
studies with correlated and multi‐group designs. Ecology, 92(11), 2049-2055.
https://doi.org/10.1890/11-0423.1
Macartney, E. L., Lagisz, M., & Nakagawa, S. (2022). The relative benefits
of environmental enrichment on learning and memory are greater when
stressed: A meta-analysis of interactions in rodents.
Neuroscience & Biobehavioral Reviews, 135, 104554.
https://doi.org/10.1016/j.neubiorev.2022.104554
}
\keyword{internal}
