% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{list_available_parameters}
\alias{list_available_parameters}
\title{List Available Parameters in Database (with parsed global_index)}
\usage{
list_available_parameters(db_path, table_name = "simulation_results")
}
\description{
List Available Parameters in Database (with parsed global_index)
}
