% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnRR.R
\name{lnRR}
\alias{lnRR}
\title{Log Response Ratio (lnRR)}
\usage{
lnRR(
  type,
  data,
  col_names = c("yi", "vi"),
  append = TRUE,
  Ctrl_mean,
  Ctrl_sd,
  Ctrl_n,
  A_mean,
  A_sd,
  A_n,
  B_mean = NULL,
  B_sd = NULL,
  B_n = NULL,
  AB_mean = NULL,
  AB_sd = NULL,
  AB_n = NULL
)
}
\arguments{
\item{type}{Type of effect size: "ind", "main", or "inter"}

\item{data}{Data frame with the data}

\item{col_names}{Vector of two strings to name the output columns for the effect size and it's sampling variance. Default is 'yi' and 'vi'.}

\item{append}{Logical. Append the results to \code{data}. Default is TRUE}

\item{Ctrl_mean}{Mean outcome from the Control treatment}

\item{Ctrl_sd}{Standard deviation from the control treatment}

\item{Ctrl_n}{Sample size from the control streatment}

\item{A_mean}{Mean outcome from the treatment}

\item{A_sd}{Standard deviation from the treatment}

\item{A_n}{Sample size from the treatment}

\item{B_mean}{Mean outcome from the B treatment}

\item{B_sd}{Standard deviation from the B treatment}

\item{B_n}{Sample size from the B treatment}

\item{AB_mean}{Mean outcome from the interaction AxB treatment}

\item{AB_sd}{Standard deviation from the interaction AxB treatment}

\item{AB_n}{Sample size from the interaction AxB treatment}
}
\description{
Log Response Ratio (lnRR)
}
\details{
\strong{Main Log Response Ratio}

The main effect of factor A is computed as:

\deqn{
lnRR_{A} = \
  \ln \left( 
    \frac{
      \bar{x}_{AB} + \bar{x}_{A}
    }{
      \bar{x}_{B} + \bar{x}_{Ctrl}
    }
  \right)
}

\deqn{
s^2_{A} = \
  \left( \frac{1} {\bar{x}_{AB} + \bar{x}_{A}} \right)^2 \
  \left(
     \frac{sd_{AB}^2} {n_{AB}} + \frac{sd_{A}^2} {n_{A}}
  \right)
  \left( \frac{1} {\bar{x}_{B} + \bar{x}_{Ctrl}} \right)^2 \
  \left(
     \frac{sd_{B}^2} {n_{B}} + \frac{sd_{Ctrl}^2} {n_{Ctrl}}
  \right)
}

\strong{Interaction Log Response Ratio}

\deqn{
lnRR_{AB} = \left( \ln \bar{x}_{AB} - \ln \bar{x}_{B} \right) -
            \left( \ln \bar{x}_{A} - \ln \bar{x}_{Ctrl} \right)
}
}
\author{
Facundo Decunta - fdecunta@agro.uba.ar
}
