% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{local_cluster_malariasim_controller}
\alias{local_cluster_malariasim_controller}
\title{Run Malaria Simulation with Local Cluster}
\usage{
local_cluster_malariasim_controller(input, reps)
}
\arguments{
\item{input}{Input parameters list}

\item{reps}{Number of replicates}
}
\value{
List of simulation results
}
\description{
Run Malaria Simulation with Local Cluster
}
