% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{local_malariasim_controller}
\alias{local_malariasim_controller}
\title{Run Malaria Simulation Locally}
\usage{
local_malariasim_controller(input, reps)
}
\arguments{
\item{input}{Input parameters list}

\item{reps}{Number of replicates}
}
\value{
List of simulation results
}
\description{
Run Malaria Simulation Locally
}
