% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnvr.R
\name{main_lnVR}
\alias{main_lnVR}
\title{Main effect: Log Variability Ratio}
\usage{
main_lnVR(Ctrl_mean, Ctrl_sd, A_mean, A_sd, B_mean, B_sd, AB_mean, AB_sd)
}
\arguments{
\item{Ctrl_sd}{Standard deviation from the control treatment}

\item{A_sd}{Standard deviation from the A treatment}

\item{B_sd}{Standard deviation from the B treatment}

\item{AB_sd}{Standard deviation from the interaction AxB treatment}

\item{Ctrl_n}{Sample size from the control streatment}

\item{A_n}{Sample size from the A treatment}

\item{B_n}{Sample size from the B treatment}

\item{AB_n}{Sample size from the interaction AxB treatment}
}
\description{
Main effect: Log Variability Ratio
}
\keyword{internal}
