% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator_functions.R
\name{prepare_input_features_schema}
\alias{prepare_input_features_schema}
\title{Prepare input features with full schema support}
\usage{
prepare_input_features_schema(df, models, window_size)
}
\description{
Prepare input features with full schema support
}
\keyword{internal}
