% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet_functions.R
\name{set_bednet_parameters}
\alias{set_bednet_parameters}
\title{Set Bednet Parameters}
\usage{
set_bednet_parameters(simparams, lhs_sample, bednet_params, baseline = TRUE)
}
\arguments{
\item{simparams}{Simulation parameters object}

\item{lhs_sample}{Single row from LHS data}

\item{bednet_params}{Bednet parameter data}

\item{baseline}{Logical, whether this is baseline (default TRUE)}
}
\value{
List with updated simparams and timesteps
}
\description{
Set Bednet Parameters
}
