% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{set_lsm_parameters}
\alias{set_lsm_parameters}
\title{Set LSM Parameters}
\usage{
set_lsm_parameters(simparams, lhs_sample)
}
\arguments{
\item{simparams}{Simulation parameters object}

\item{lhs_sample}{Single row from LHS data}
}
\value{
List with updated simparams and timesteps
}
\description{
Set LSM Parameters
}
