% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{set_mosquito_parameters}
\alias{set_mosquito_parameters}
\title{Set Mosquito Parameters}
\usage{
set_mosquito_parameters(lhs_sample)
}
\arguments{
\item{lhs_sample}{Single row from LHS data}
}
\value{
List of mosquito parameters
}
\description{
Set Mosquito Parameters
}
