% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{set_seasonality}
\alias{set_seasonality}
\title{Set Seasonality Parameters}
\usage{
set_seasonality(lhs_sample)
}
\arguments{
\item{lhs_sample}{Single row from LHS data}
}
\value{
List with g0, g, and h parameters
}
\description{
Set Seasonality Parameters
}
