% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnrr.R
\name{simple_lnRR}
\alias{simple_lnRR}
\title{Simple effect: log response ratio}
\usage{
simple_lnRR(Ctrl_mean, Ctrl_sd, Ctrl_n, A_mean, A_sd, A_n)
}
\arguments{
\item{Ctrl_mean}{Mean outcome from the Control treatment}

\item{Ctrl_sd}{Standard deviation from the control treatment}

\item{Ctrl_n}{Sample size from the control streatment}

\item{A_mean}{Mean outcome from the treatment}

\item{A_sd}{Standard deviation from the treatment}

\item{A_n}{Sample size from the treatment}
}
\description{
Computes the log of the response ratio
}
\keyword{internal}
