% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMD_wrappers.R
\name{SMD_ind}
\alias{SMD_ind}
\title{Simple effect: Standardized Mean Difference}
\usage{
SMD_ind(
  data,
  col_names = c("yi", "vi"),
  append = TRUE,
  hedges_correction = TRUE,
  Ctrl_mean,
  Ctrl_sd,
  Ctrl_n,
  A_mean,
  A_sd,
  A_n
)
}
\arguments{
\item{data}{Data frame containing the variables used.}

\item{col_names}{Vector of two strings to name the output columns for the effect size and it's sampling variance. Default is 'yi' and 'vi'.}

\item{append}{Logical. Append the results to \code{data}. Default is TRUE}

\item{hedges_correction}{Boolean. If TRUE correct for small-sample bias. Default is TRUE.}

\item{Ctrl_mean}{Mean outcome from the Control treatment}

\item{Ctrl_sd}{Standard deviation from the control treatment}

\item{Ctrl_n}{Sample size from the control streatment}

\item{A_mean}{Mean outcome from the experimental treatment}

\item{A_sd}{Standard deviation from the experimental treatment}

\item{A_n}{Sample size from the experimental treatment}
}
\description{
Computes the individual or simple effect of Factor A over the Control.
}
\details{
It is the classic Standardized Mean Difference (SMD), which can also be computed
with metafor's \code{escalc()} function using \code{measure = "SMD"}.

See the package vignette for a detailed description of the formula.
}
\references{
Gurevitch, J., Morrison, J. A., & Hedges, L. V. (2000). The interaction
between competition and predation: a meta-analysis of field experiments.
The American Naturalist, 155(4), 435-453.

Morris, W. F., Hufbauer, R. A., Agrawal, A. A., Bever, J. D., Borowicz, V. A.,
Gilbert, G. S., ... & Vázquez, D. P. (2007). Direct and interactive
effects of enemies and mutualists on plant performance: a meta‐analysis.
Ecology, 88(4), 1021-1029. https://doi.org/10.1890/06-0442
}
