% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINTweb_parser.R
\name{available_net_types}
\alias{available_net_types}
\title{Return the vector of net types contained in an ITN-parameter file}
\usage{
available_net_types(itn_params_path = NULL)
}
\arguments{
\item{itn_params_path}{Path to an RDS created by MINTer}
}
\description{
Return the vector of net types contained in an ITN-parameter file
}
