% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_effsize.R
\name{compute_effsize}
\alias{compute_effsize}
\title{This function should be the main API}
\usage{
compute_effsize(
  type,
  effsize,
  colname = c("yi", "vi"),
  data,
  Ctrl_mean,
  Ctrl_sd,
  Ctrl_n,
  A_mean,
  A_sd,
  A_n,
  B_mean,
  B_sd,
  B_n,
  AB_mean,
  AB_sd,
  AB_n
)
}
\arguments{
\item{type}{Type of effect size: individual, main or interaction}

\item{effsize}{The effect size to compute: SMD, lnRR, lnVR or lnCVR}

\item{colname}{List with two strings. Names of the output columns.
By default uses the same as metafor: yi, vi.}

\item{data}{Data frame with the data.}

\item{Ctrl_mean}{Mean outcome from the Control treatment.}

\item{Ctrl_sd}{Standard deviation from the control treatment.}

\item{Ctrl_n}{Sample size from the control streatment.}

\item{A_mean}{Mean outcome from the A treatment.}

\item{A_sd}{Standard deviation from the A treatment.}

\item{A_n}{Sample size from the A treatment.}

\item{B_mean}{Mean outcome from the B treatment.}

\item{B_sd}{Standard deviation from the B treatment.}

\item{B_n}{Sample size from the B treatment.}

\item{AB_mean}{Mean outcome from the interaction AxB treatment.}

\item{AB_sd}{Standard deviation from the interaction AxB treatment.}

\item{AB_n}{Sample size from the interaction AxB treatment.}
}
\description{
Needs for each effect size:
}
