% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator_main.R
\name{convert_db_results_to_dataframe}
\alias{convert_db_results_to_dataframe}
\title{Convert Database Results to Standard Dataframe}
\usage{
convert_db_results_to_dataframe(raw_results, predictor, model_types)
}
\arguments{
\item{raw_results}{Results from run_emulator_db}

\item{predictor}{"prevalence" or "cases"}

\item{model_types}{Vector of model types used}
}
\value{
Data frame with standardized format
}
\description{
Convert Database Results to Standard Dataframe
}
\keyword{internal}
