% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{create_scenario_plots}
\alias{create_scenario_plots}
\title{Create Scenario Plots (Helper Function)}
\usage{
create_scenario_plots(
  predictions,
  models,
  model_types,
  window_size = 14,
  output_dir = NULL
)
}
\arguments{
\item{predictions}{List of predictions from generate_scenario_predictions}

\item{models}{List from load_emulator_models()}

\item{model_types}{Vector of model types that were used}

\item{window_size}{Window size for time scaling}

\item{output_dir}{Directory to save plots}
}
\value{
List of ggplot objects
}
\description{
Create Scenario Plots (Helper Function)
}
\keyword{internal}
