% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINTweb_parser.R
\name{define_bednet_types}
\alias{define_bednet_types}
\title{Define bednet types and create spline fits}
\usage{
define_bednet_types(net_types, itn_params_path = NULL, strict = FALSE)
}
\arguments{
\item{itn_params_path}{Path to RDS file containing ITN parameters. If NULL, uses bundled data.}

\item{strict}{Allows for future expansion of dataset (defaul FALSE to use only available nets in RDS)}

\item{NET_TYPES}{Character vector of net types to process}
}
\value{
Named list of smooth.spline objects, one for each net type
}
\description{
Define bednet types and create spline fits
}
