% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMD.R
\name{.j_correction}
\alias{.j_correction}
\title{Correction for small-sample bias}
\usage{
.j_correction(m)
}
\arguments{
\item{m}{Degrees of freedom. They change between individual and main effects.}
}
\description{
Small-sample bias correction for Hedges' g
}
\keyword{internal}
