% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnRR.R
\name{.main_lnRR_Nakagawa}
\alias{.main_lnRR_Nakagawa}
\title{Main Log Response Ratio: method by Nakagawa}
\usage{
.main_lnRR_Nakagawa(
  Ctrl_mean,
  Ctrl_sd,
  Ctrl_n,
  A_mean,
  A_sd,
  A_n,
  B_mean,
  B_sd,
  B_n,
  AB_mean,
  AB_sd,
  AB_n
)
}
\arguments{
\item{Ctrl_mean}{Mean outcome from the Control treatment}

\item{Ctrl_sd}{Standard deviation from the control treatment}

\item{Ctrl_n}{Sample size from the control streatment}

\item{A_mean}{Mean outcome from the A treatment}

\item{A_sd}{Standard deviation from the A treatment}

\item{A_n}{Sample size from the A treatment}

\item{B_mean}{Mean outcome from the B treatment}

\item{B_sd}{Standard deviation from the B treatment}

\item{B_n}{Sample size from the B treatment}

\item{AB_mean}{Mean outcome from the interaction AxB treatment}

\item{AB_sd}{Standard deviation from the interaction AxB treatment}

\item{AB_n}{Sample size from the interaction AxB treatment}
}
\description{
Computes the main effect of a treatment in a 2-by-2 factorial design using
the method proposed by Nakagawa (in prep.)
}
\details{
Given a full factorial design with factors A x B
The main effect of A is quantified as the log of the ratio of the average
outcome in the two treatments where A is present. That is, treatments A-and-B, A-and-Control.
}
\references{
Not published yet.
}
\keyword{internal}
