% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{.pooled_sd}
\alias{.pooled_sd}
\title{Pooled Standard Deviation for SMD in factorial experiments}
\usage{
.pooled_sd(Ctrl_sd, Ctrl_n, A_sd, A_n, B_sd, B_n, AB_n, AB_sd)
}
\description{
TODO:
}
\references{
Morris, W. F., Hufbauer, R. A., Agrawal, A. A., Bever, J. D., Borowicz, V. A.,
Gilbert, G. S., ... & Vázquez, D. P. (2007). Direct and interactive
effects of enemies and mutualists on plant performance: a meta‐analysis.
Ecology, 88(4), 1021-1029. https://doi.org/10.1890/06-0442
}
\keyword{internal}
