% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnVR.R
\name{.simple_lnVR}
\alias{.simple_lnVR}
\title{Simple Log Variability ratio}
\usage{
.simple_lnVR(Ctrl_sd, Ctrl_n, X_sd, X_n)
}
\arguments{
\item{Ctrl_sd}{Standard deviation from the control treatment}

\item{Ctrl_n}{Sample size from the control streatment}

\item{A_sd}{Standard deviation from the A treatment}

\item{A_n}{Sample size from the A treatment}

\item{B_sd}{Standard deviation from the B treatment}

\item{B_n}{Sample size from the B treatment}

\item{AB_sd}{Standard deviation from the interaction AxB treatment}

\item{AB_n}{Sample size from the interaction AxB treatment}
}
\description{
Simple Log Variability ratio
}
\references{
Nakagawa, S., Poulin, R., Mengersen, K., Reinhold, K., Engqvist,
L., Lagisz, M., & Senior, A. M. (2015). Meta‐analysis of variation:
ecological and evolutionary applications and beyond. Methods in
Ecology and Evolution, 6(2), 143-152.
}
\keyword{internal}
