% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMD.R
\name{.simple_SMD}
\alias{.simple_SMD}
\title{Simple effect: Standardized Mean Difference}
\usage{
.simple_SMD(
  Ctrl_mean,
  Ctrl_sd,
  Ctrl_n,
  A_mean,
  A_sd,
  A_n,
  hedges_correction = TRUE
)
}
\arguments{
\item{Ctrl_mean}{Mean outcome from the Control treatment}

\item{Ctrl_sd}{Standard deviation from the control treatment}

\item{Ctrl_n}{Sample size from the control streatment}

\item{A_mean}{Mean outcome from experimental treatment}

\item{A_sd}{Standard deviation from experimental treatment}

\item{A_n}{Sample size from experimental treatment}

\item{hedges_correction}{Boolean. If TRUE correct for small-sample bias. Default is TRUE.}
}
\description{
Simple effect: Standardized Mean Difference
}
\references{
Gurevitch, J., Morrison, J. A., & Hedges, L. V. (2000). The interaction
between competition and predation: a meta-analysis of field experiments.
The American Naturalist, 155(4), 435-453.

Morris, W. F., Hufbauer, R. A., Agrawal, A. A., Bever, J. D., Borowicz, V. A.,
Gilbert, G. S., ... & Vázquez, D. P. (2007). Direct and interactive
effects of enemies and mutualists on plant performance: a meta‐analysis.
Ecology, 88(4), 1021-1029. https://doi.org/10.1890/06-0442
}
\keyword{internal}
