% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator_main.R
\name{generate_scenario_predictions_batched}
\alias{generate_scenario_predictions_batched}
\title{Generate Scenario Predictions with Batching}
\usage{
generate_scenario_predictions_batched(
  scenarios,
  models,
  model_types = c("GRU", "LSTM"),
  time_steps = 2190,
  window_size = 14,
  benchmark = FALSE
)
}
\arguments{
\item{scenarios}{Data frame with scenario parameters}

\item{models}{List from load_emulator_models()}

\item{model_types}{Vector of model types}

\item{time_steps}{Number of time steps}

\item{window_size}{Window size for rolling average}

\item{benchmark}{Track timing details}
}
\value{
List with predictions for each scenario
}
\description{
Generate Scenario Predictions with Batching
}
\keyword{internal}
