% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_cached_model}
\alias{get_cached_model}
\title{Get Cached Models or Load if Missing}
\usage{
get_cached_model(model_type)
}
\arguments{
\item{model_type}{"eir_models", "nn_prevalence", or "nn_cases"}
}
\value{
Cached model object (may trigger a load if absent)
}
\description{
Get Cached Models or Load if Missing
}
