% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{initialize_python}
\alias{initialize_python}
\title{Initialize Python Dependencies}
\usage{
initialize_python()
}
\description{
This function initializes the Python dependencies required for the emulator.
It is called automatically when needed, but can be called manually to
pre-load the dependencies.
}
