% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_effsize.R
\name{inter_effsize}
\alias{inter_effsize}
\title{Compute effect sizes from 2x2 factorial experiment}
\usage{
inter_effsize(
  effsize,
  Ctrl_mean,
  Ctrl_sd,
  Ctrl_n,
  A_mean,
  A_sd,
  A_n,
  B_mean,
  B_sd,
  B_n,
  AB_mean,
  AB_sd,
  AB_n
)
}
\description{
Compute effect sizes from 2x2 factorial experiment
}
