% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnVR.R
\name{lnVR}
\alias{lnVR}
\title{Log of Variation Ratio (lnVR)}
\usage{
lnVR(
  type,
  data,
  col_names = c("yi", "vi"),
  append = TRUE,
  Ctrl_sd,
  Ctrl_n,
  A_sd,
  A_n,
  B_sd = NULL,
  B_n = NULL,
  AB_sd = NULL,
  AB_n = NULL
)
}
\arguments{
\item{type}{Type of effect size: "ind", "main", or "inter"}

\item{data}{Data frame containing the variables used.}

\item{col_names}{Vector of two strings to name the output columns for the effect size and it's sampling variance. Default is 'yi' and 'vi'.}

\item{append}{Logical. Append the results to \code{data}. Default is TRUE}

\item{Ctrl_sd}{Standard deviation from the control treatment}

\item{Ctrl_n}{Sample size from the control streatment}

\item{A_sd}{Standard deviation from the treatment}

\item{A_n}{Sample size from the treatment}

\item{B_sd}{Standard deviation from the B treatment}

\item{B_n}{Sample size from the B treatment}

\item{AB_sd}{Standard deviation from the interaction AxB treatment}

\item{AB_n}{Sample size from the interaction AxB treatment}
}
\description{
Log of Variation Ratio (lnVR)
}
\author{
Facundo Decunta - fdecunta@agro.uba.ar
}
