% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnVR_wrappers.R
\name{lnVR_ind}
\alias{lnVR_ind}
\title{Individual effect: Log of Variability Ratio}
\usage{
lnVR_ind(
  data,
  col_names = c("yi", "vi"),
  append = TRUE,
  Ctrl_sd,
  Ctrl_n,
  A_sd,
  A_n
)
}
\arguments{
\item{data}{Data frame containing the variables used.}

\item{col_names}{Vector of two strings to name the output columns for the effect size and it's sampling variance. Default is 'yi' and 'vi'.}

\item{append}{Logical. Append the results to \code{data}. Default is TRUE}

\item{Ctrl_sd}{Standard deviation from the control treatment}

\item{Ctrl_n}{Sample size from the control streatment}

\item{A_sd}{Standard deviation from the treatment}

\item{A_n}{Sample size from the treatment}
}
\description{
Computes the Log of the Variability Ratio between
a Factor A and the Control treatment in factorial experiments.
}
\details{
See the package vignette for a detailed description of the formula.
}
\references{
Nakagawa, S., Poulin, R., Mengersen, K., Reinhold, K., Engqvist,
L., Lagisz, M., & Senior, A. M. (2015). Meta‐analysis of variation:
ecological and evolutionary applications and beyond. Methods in
Ecology and Evolution, 6(2), 143-152.
}
\author{
Facundo Decunta - fdecunta@agro.uba.ar
}
