% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINTer-package.R
\docType{package}
\name{MINTer-package}
\alias{MINTer}
\alias{MINTer-package}
\title{MINTer: Malaria INTervention Emulator in R}
\description{
A comprehensive package for malaria intervention simulation and neural network-based emulation. Combines malariasimulation with trained GRU/LSTM models for rapid prediction of intervention outcomes.
}
\author{
\strong{Maintainer}: Cosmo Santoni \email{cosmo.santoni@imperial.ac.uk}

}
\keyword{internal}
