% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINTweb_controller.R
\name{run_mint_scenarios}
\alias{run_mint_scenarios}
\title{Run MINT Scenarios for Malaria Intervention Analysis}
\usage{
run_mint_scenarios(
  res_use,
  res_future,
  py_only,
  py_pbo,
  py_pyrrole,
  py_ppf,
  prev_vec,
  Q0_vec,
  phi_vec,
  season_vec,
  routine_vec,
  irs_vec,
  irs_future_vec,
  lsm_vec,
  eir_models = "xgboost",
  prevalence_models = "LSTM",
  predictor = c("prevalence", "cases")
)
}
\arguments{
\item{res_use}{Numeric vector of current resistance levels}

\item{res_future}{Numeric vector of future resistance levels after next campaign}

\item{py_only}{Numeric vector of pyrethroid-only net proportions}

\item{py_pbo}{Numeric vector of pyrethroid-PBO net proportions}

\item{py_pyrrole}{Numeric vector of pyrethroid-pyrrole net proportions}

\item{py_ppf}{Numeric vector of pyrethroid-PPF net proportions}

\item{prev_vec}{Numeric vector of malaria prevalence levels}

\item{Q0_vec}{Numeric vector of Q0 values}

\item{phi_vec}{Numeric vector of phi bednet values}

\item{season_vec}{Numeric vector of seasonality indicators}

\item{routine_vec}{Numeric vector of routine treatment indicators}

\item{irs_vec}{Numeric vector of current IRS coverage}

\item{irs_future_vec}{Numeric vector of future IRS coverage}

\item{lsm_vec}{Numeric vector of LSM coverage}

\item{eir_models}{Character vector of models to use for EIR prediction (default: "xgboost")}

\item{prevalence_models}{Character vector of models for prevalence prediction (default: "LSTM")}

\item{predictor}{Character vector of predictors to run (default: c("prevalence", "cases"))}
}
\value{
List of prevalence and cases predictions
}
\description{
Run MINT Scenarios for Malaria Intervention Analysis
}
