% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINTweb_controller.R
\name{run_mint_scenarios}
\alias{run_mint_scenarios}
\title{Run MINT Scenarios for Malaria Intervention Analysis}
\usage{
run_mint_scenarios(
  res_use,
  res_future = NULL,
  py_only,
  py_pbo,
  py_pyrrole,
  py_ppf,
  net_type_future = NULL,
  itn_future = NULL,
  prev,
  Q0,
  phi,
  season,
  routine,
  irs,
  irs_future,
  lsm,
  eir_models = "xgboost",
  prevalence_models = "LSTM",
  predictor = c("prevalence", "cases"),
  year_start = 2,
  year_end = 5,
  scenario_tag = NULL,
  cull_prevalence = c(2, 7),
  output_time = FALSE
)
}
\arguments{
\item{res_use}{Numeric vector of current resistance levels}

\item{res_future}{Numeric vector of future resistance levels after next campaign}

\item{py_only}{Numeric vector of pyrethroid-only net proportions}

\item{py_pbo}{Numeric vector of pyrethroid-PBO net proportions}

\item{py_pyrrole}{Numeric vector of pyrethroid-pyrrole net proportions}

\item{py_ppf}{Numeric vector of pyrethroid-PPF net proportions}

\item{net_type_future}{Character vector of future net choices per scenario
(allowed: "py_only","py_pbo","py_pyrrole","py_ppf"); use NA for baseline.}

\item{itn_future}{Numeric vector (0–1) of future ITN coverage per scenario; use NA for baseline.}

\item{prev}{Numeric vector of malaria prevalence levels}

\item{Q0}{Numeric vector of Q0 values}

\item{phi}{Numeric vector of phi bednet values}

\item{season}{Numeric vector of seasonality indicators}

\item{routine}{Numeric vector of routine treatment indicators}

\item{irs}{Numeric vector of current IRS coverage}

\item{irs_future}{Numeric vector of future IRS coverage}

\item{lsm}{Numeric vector of LSM coverage}

\item{eir_models}{Character vector of models to use for EIR prediction (default: "xgboost")}

\item{prevalence_models}{Character vector of models for prevalence prediction (default: "LSTM")}

\item{predictor}{Character vector of predictors to run (default: c("prevalence", "cases"))}

\item{year_start, year_end}{Integers for case-year range}

\item{scenario_tag}{Optional character vector of scenario identifiers. If provided,
it must have length equal to the number of scenarios. Defaults to "Scenario1", "Scenario2", ...}

\item{cull_prevalence}{Optional integer vector to remove the first X years worth of simulation and aggregate}

\item{output_time}{Optional bool to cull absolute time tracking from output objects}
}
\value{
List of prevalence and cases predictions
}
\description{
Run MINT Scenarios for Malaria Intervention Analysis
}
