% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{simple_SMD}
\alias{simple_SMD}
\title{Simple effect: Standardized Mean Difference Hedges' g}
\usage{
simple_SMD(Ctrl_mean, Ctrl_sd, Ctrl_n, X_mean, X_sd, X_n, pooled_sd)
}
\arguments{
\item{Ctrl_mean}{Mean outcome from the Control treatment}

\item{Ctrl_sd}{Standard deviation from the control treatment}

\item{Ctrl_n}{Sample size from the control streatment}

\item{X_mean}{Mean outcome from treatment}

\item{X_sd}{Standard deviation from treatment}

\item{X_n}{Sample size from treatment}

\item{pooled_sd}{Pooled standard deviation}
}
\description{
Simple effect: Standardized Mean Difference Hedges' g
}
\keyword{internal}
